refactor: categories를 data에서 constants로 이동
